modularize prescription model by splitting original 2 files into 6 separate files for better structure and maintainability
