+++ backend/routes/index.js
@@ -31,7 +31,7 @@ const UserRoutes  = require('../models/user/user.route');
 const AuthRoutes = require('./auth');
 // const DoctorRoutes = require('../modules/doctor/doctor.route');
 // const AppointmentRoutes = require('../modules/appointment/appointment.route');
-// const PrescriptionRoutes = require('../modules/prescription/prescription.route');
+const PrescriptionRoutes = require('../models/prescription/prescription.route');
 // const MedicalRecordRoutes = require('../modules/medicalRecord/medicalRecord.route');
 const LabTestRoutes = require('../models/labTest/labTest.route');
 // const PaymentRoutes = require('../modules/payment/payment.route');
@@ -47,7 +47,7 @@ router.use('/user', UserRoutes);
 router.use('/auth', AuthRoutes);
 // router.use('/doctors', DoctorRoutes);
 // router.use('/appointments', AppointmentRoutes);
-// router.use('/prescriptions', PrescriptionRoutes);
+router.use('/prescriptions', PrescriptionRoutes);
 // router.use('/medical-records', MedicalRecordRoutes);
 router.use('/lab-tests', LabTestRoutes);
 // router.use('/payments', PaymentRoutes);
